# Improve performance for large (oversegmentation) meshes