@daniel-wer 2023-05-15
> Since you implemented the merging of meshes in one chunk, the issue https://github.com/scalableminds/webknossos/issues/6519 just got even easier. I think, it would only be a matter of adding a line à la geometry = mergeVertices(geometry);.
> 
> Do you feel like giving this a try to see whether:
> 
> - this reduces visible mesh borders
> - the performance hit is negligible

I tested that, see this log for some time measurements and how many vertices were removed for some sample chunks for lods 0 and 4:

```
console.js:13 Before 9045 lod 0
isosurface_saga.ts:863 simplification: 50.305908203125 ms
console.js:13 After 8859 lod 0
console.js:13 Before 7085 lod 0
isosurface_saga.ts:863 simplification: 29.9140625 ms
console.js:13 After 6932 lod 0
console.js:13 Before 12333 lod 0
isosurface_saga.ts:863 simplification: 73.60693359375 ms
console.js:13 After 11944 lod 0
console.js:13 Before 12137 lod 0
isosurface_saga.ts:863 simplification: 115.741943359375 ms
console.js:13 After 11763 lod 0
console.js:13 Before 12848 lod 4
isosurface_saga.ts:863 simplification: 84.89111328125 ms
console.js:13 After 10499 lod 4
console.js:13 Before 13198 lod 4
isosurface_saga.ts:863 simplification: 62.898193359375 ms
console.js:13 After 12019 lod 4
console.js:13 Before 13172 lod 4
isosurface_saga.ts:863 simplification: 62.249755859375 ms
console.js:13 After 11403 lod 4
console.js:13 Before 10787 lod 4
isosurface_saga.ts:863 simplification: 88.73193359375 ms
console.js:13 After 9325 lod 4
console.js:13 Before 11184 lod 4
isosurface_saga.ts:863 simplification: 55.5029296875 ms
console.js:13 After 9816 lod 4
```

For the large mesh I'm testing with, this would be called thousands of times so the performance impact would be noticeable I'd say. For the in-chunk borders, I don't see any difference :/

Before:
![chunk_borders_before](https://github.com/scalableminds/webknossos/assets/1702075/dfba6076-499a-4ca4-9f2e-d75f2aecd8ec)
After:
![chunk_borders_after](https://github.com/scalableminds/webknossos/assets/1702075/07edd5e5-89f8-44c4-befd-e2604e3cfd83)